v10.0.2
This version is compatible with 10.0.1 lattice
This version was tested and debuged against elder v8.0.0.a1 to find the error in base mapping. A new base mapping has been designed also to be sure the calculation is correct. The real error was in I0 and I1 modified Bessel functions and in the wrong assignment of coordinates in the start vectors for the orthogonal tracks in lattice:cs_traj.